Feat/18-report: 추출된 키워드 조회 후 보고서 생성 파이프라인 구축
