+++ src/main/java/com/likelion/cleopatra/domain/report/entity/Report.java
@@ -1,4 +1,3 @@
-// src/main/java/com/likelion/cleopatra/domain/report/entity/Report.java
 package com.likelion.cleopatra.domain.report.entity;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -12,14 +11,14 @@ import com.likelion.cleopatra.global.common.enums.keyword.Secondary;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-/** 실무 포인트
+/**
  * - 조인 없음. 섹션별 JSON 컬럼으로 분할 저장.
  * - dev(H2)에서 json 타입이 깨지면 columnDefinition="TEXT"로 변경.
  * - 추후 인덱싱 필요 시 가상컬럼+인덱스 추가.
  */
+
 @Getter
 @Builder
 @AllArgsConstructor
@@ -94,7 +93,7 @@ public class Report {
 
                 // 섹터별 Json 형식
                 .descriptionSummaryJson(write(om, total.getDescriptionSummary()))
-                .keywordsJson(write(om, total.getKeywords()))
+                .keywordsJson(write(om, total.getKeywordReportRes().getKeywords()))
                 .populationJson(write(om, total.getPopulation()))
                 .priceJson(write(om, total.getPrice()))
                 .incomeConsumptionJson(write(om, total.getIncomeConsumption()))
